Split costmap calculation into a class.
